Enabling panic condition for Widening/Narrowing
NOTE: There is a very initial support for widening convert from int32 to fp64. Basic support for all widening and narrowing conversions will be addded soon, but limiting to MVL/2 in case of widining, ince we are not implementing register grouping. This will allow to convert between int32,int64,fp32 and fp64.