arch-riscv: Change the nature of vsetvl* family of insts
- The vsetvl* insts, apart from the spec, also update the vl and
vtype stored in the next pc PCState.
- The vsetvl* insts are marked as control instructions and behave
like a branch instruction. Upon the execution stage, the values of
vl and vtype will be updated, and the updated values will be
compared to the initial values. If they are different, the
pipeline will be flushed. The next instruction will get the PCState
from this instruction, so it will have the updated vl and vtype.

Change-Id: I8d988fd8ca833021908694bd1713cb3963bdfa3c
Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>